refactor(editor): split vim engine into per-domain extensions
